requirements.txt
streamlit==1.12.0
Pillow==9.3.0
altair==4.0.1
plotly-express==0.4.1
plotly==5.11.0
pip==23.2.1
streamlit-option-menu==0.3.6